fix(audit): pass 1-4 — 12 HIGH, 18 MEDIUM, ~55 LOW + NITs
PR #544 fresh re-audit (82 files, ~63K lines, 8 partitions). Four passes
to clean: 12 HIGH (P5×3 + P6×2 + P7×3 + P4×2 + P2×2), 18 MEDIUM, ~55
LOW/NIT. Cumulative across passes 1-4.

Highlights:
* Quarantine corrupt windowrules.json before rebuild (data loss).
* Log non-string disable values during v3→v4 migration.
* Remove dead isContextDisabled fallback in updateAutotileScreens.
* Wire WTA::windowClosedNotification → WDA::handleWindowClosed (replaces
  dead D-Bus method); plain public member (not Q_SLOTS) to keep off the
  bus surface; fixed Pass-1 infinite recursion regression.
* qFatal instead of Q_ASSERT_X(false) in enginewiring.cpp.
* Clear animation rule cache on windowClass/desktopFileName change.
* Cache SetOpacity resolve across prePaintWindow/paintWindow (per-frame
  cache cleared at postPaintScreen + windowDeleted).
* Use isWindowFloating (appId-aware) inside setWindowFloating early-return.
* Gate assignWindowToZones markDirty on zoneChanged || screenChanged ||
  desktopChanged.
* Cache texCoord handedness on ShaderTransition.
* allModes() in settingscontroller_session prune loops (picks up Scrolling).
* setContextLocked rejects (desktop=0, non-empty activity) at boundary.
* PZ_STORE_SET_BOOL/COLOR/STRING macros now use pre+post read-back.
* Gate purge/clear paths on isPureAssignmentRule to preserve non-assignment
  actions (SetOpacity, OverrideAnimation*) on rule mutation.

Carry-overs descoped: file-size violations (16 files exceed 800-line cap;
flagged in 14+ prior passes, no commit splits them; descoped via PR note).

211/211 tests pass.

+++ src/daemon/contextresolverwiring.cpp
@@ -102,11 +102,13 @@ bool DaemonSettingsGateAdapter::isContextLocked(PhosphorZones::AssignmentEntry::
     if (!m_settings)
         return false;
     // Fold the Mode into the screen-key the way every pre-migration call
-    // site did (`Utils::contextLockKey(static_cast<int>(mode), screenId)`).
-    // The lock store is keyed by composite "mode:screenId" strings so the
-    // Mode axis stays implicit on the wire format — changing that is a
-    // schema migration, not an adapter concern.
-    return m_settings->isContextLocked(Utils::contextLockKey(mode, screenId), desktop, activity);
+    // site did. Explicit cast over the unscoped-enum implicit conversion:
+    // the wire format takes an int, and writing the cast keeps the call
+    // safe if `AssignmentEntry::Mode` ever migrates to `enum class`.
+    // The lock store is keyed by composite "mode:screenId" strings so
+    // the Mode axis stays implicit on the wire format — changing that
+    // is a schema migration, not an adapter concern.
+    return m_settings->isContextLocked(Utils::contextLockKey(static_cast<int>(mode), screenId), desktop, activity);
 }
 
 } // namespace PlasmaZones
